+++ frontend/ios/App/App/NativeTabBarPlugin.m
@@ -1,0 +1,10 @@
+#import <Foundation/Foundation.h>
+#import <Capacitor/Capacitor.h>
+
+CAP_PLUGIN(NativeTabBarPlugin, "NativeTabBar",
+    CAP_PLUGIN_METHOD(configure, CAPPluginReturnPromise);
+    CAP_PLUGIN_METHOD(setSelected, CAPPluginReturnPromise);
+    CAP_PLUGIN_METHOD(setAppearance, CAPPluginReturnPromise);
+    CAP_PLUGIN_METHOD(show, CAPPluginReturnPromise);
+    CAP_PLUGIN_METHOD(hide, CAPPluginReturnPromise);
+)